backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
openai

